+++ src/Platforms/Android/AndroidWindow.cpp
@@ -1,5 +1,5 @@
-#include "SDL_video.h"
 #ifdef TB_PLATFORM_ANDROID
+#include "SDL_video.h"
 
 #include "Platforms/Android/AndroidWindow.h"
 #include "tbpch.h"
@@ -41,6 +41,10 @@ void AndroidWindow::Init(const WindowProps& props)
     m_Data.Title = props.Title;
     m_Data.Width = props.Width;
     m_Data.Height = props.Height;
+    m_Data.MinWidth = props.MinWidth;
+    m_Data.MinHeight = props.MinHeight;
+    m_Data.Resizeable = props.Resizeable;
+    m_Data.VSync = props.VSync;
 
     TB_CORE_INFO("Creating window {0} ({1}, {2})", props.Title, props.Width, props.Height);
     SDL_LogSetAllPriority(SDL_LOG_PRIORITY_VERBOSE);
@@ -86,7 +90,9 @@ void AndroidWindow::Init(const WindowProps& props)
     m_Context->Init();
 
     SDL_SetWindowData(m_Window, "WindowData", &m_Data);
-    SetVSync(false);
+    SetVSync(props.VSync);
+    SetResizable(props.Resizeable);
+    SetMinSize(props.MinWidth, props.MinHeight);
 
     Input::Init();
 }
@@ -196,6 +202,28 @@ void AndroidWindow::SetVSync(bool enabled)
 bool AndroidWindow::IsVSync() const
 {
     return m_Data.VSync;
+}
+
+void AndroidWindow::SetResizable(bool enabled)
+{
+
+    if (enabled)
+        SDL_SetWindowResizable(m_Window, SDL_TRUE);
+    else
+        SDL_SetWindowResizable(m_Window, SDL_FALSE);
+    m_Data.Resizeable = enabled;
+}
+
+bool AndroidWindow::GetResizable() const
+{
+    return m_Data.Resizeable;
+}
+
+void AndroidWindow::SetMinSize(uint32_t minWidth, uint32_t minHeight)
+{
+    SDL_SetWindowMinimumSize(m_Window, minWidth, minHeight);
+    m_Data.MinWidth = minWidth;
+    m_Data.MinHeight = minHeight;
 }
 }