+++ src/Platforms/Linux/LinuxWindow.cpp
@@ -39,13 +39,17 @@ void LinuxWindow::Init(const WindowProps& props)
     m_Data.Title = props.Title;
     m_Data.Width = props.Width;
     m_Data.Height = props.Height;
+    m_Data.MinWidth = props.MinWidth;
+    m_Data.MinHeight = props.MinHeight;
+    m_Data.Resizeable = props.Resizeable;
+    m_Data.VSync = props.VSync;
 
     TB_CORE_INFO("Creating window {0} ({1}, {2})", props.Title, props.Width, props.Height);
 
     if (s_SDLWindowCount == 0) {
         TB_PROFILE_SCOPE_NAME("SDL Init");
         int success = SDL_Init(SDL_INIT_VIDEO | SDL_INIT_JOYSTICK);
-        TB_CORE_ASSERT_TAGGED(success + 1, "Could not initialize SDL2!");
+        TB_CORE_ASSERT_TAGGED(success, "Could not initialize SDL2!");
     }
 
     {
@@ -82,7 +86,9 @@ void LinuxWindow::Init(const WindowProps& props)
     m_Context->Init();
 
     SDL_SetWindowData(m_Window, "WindowData", &m_Data);
-    SetVSync(false);
+    SetVSync(props.VSync);
+    SetResizable(props.Resizeable);
+    SetMinSize(props.MinWidth, props.MinHeight);
 
     Input::Init();
 }
@@ -192,6 +198,28 @@ void LinuxWindow::SetVSync(bool enabled)
 bool LinuxWindow::IsVSync() const
 {
     return m_Data.VSync;
+}
+
+void LinuxWindow::SetResizable(bool enabled)
+{
+
+    if (enabled)
+        SDL_SetWindowResizable(m_Window, SDL_TRUE);
+    else
+        SDL_SetWindowResizable(m_Window, SDL_FALSE);
+    m_Data.Resizeable = enabled;
+}
+
+bool LinuxWindow::GetResizable() const
+{
+    return m_Data.Resizeable;
+}
+
+void LinuxWindow::SetMinSize(uint32_t minWidth, uint32_t minHeight)
+{
+    SDL_SetWindowMinimumSize(m_Window, minWidth, minHeight);
+    m_Data.MinWidth = minWidth;
+    m_Data.MinHeight = minHeight;
 }
 }